project/requirements.txt
fastapi
uvicorn[standard]
pydantic
mcp
google-generativeai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
chromadb
pypdf2
sentence-transformers
